split perfAnalysis to controller and service
